Moving actions into right files
